+++ src/proto_demo_noops/lib_protocol/main.ml
@@ -43,33 +43,21 @@ type block_header = {
 
 type block_header_metadata = unit
 
-let block_header_metadata_encoding_with_legacy_attestation_name =
-  Data_encoding.unit
-
 let block_header_metadata_encoding = Data_encoding.unit
 
 type operation_data = unit
 
 let operation_data_encoding = Data_encoding.unit
 
-let operation_data_encoding_with_legacy_attestation_name =
-  operation_data_encoding
-
 type operation_receipt = unit
 
 let operation_receipt_encoding = Data_encoding.unit
-
-let operation_receipt_encoding_with_legacy_attestation_name =
-  operation_receipt_encoding
 
 let operation_data_and_receipt_encoding =
   Data_encoding.conv
     (function ((), ()) -> ())
     (fun () -> ((), ()))
     Data_encoding.unit
-
-let operation_data_and_receipt_encoding_with_legacy_attestation_name =
-  operation_data_and_receipt_encoding
 
 type operation = {
   shell : Operation.shell_header;